requirements.txt
discord.py
python-dotenv
aiohttp
aiosqlite
openai
requests
